refactor: final structure change
