+++ CMakeLists.txt
@@ -34,8 +34,11 @@ add_library(${PROJECT_NAME} SHARED
   "src/${PROJECT_NAME}/control/controller/packet/protocol_1/status/status_packet.cpp"
   "src/${PROJECT_NAME}/control/controller/packet/protocol_1/utils/word.cpp"
   "src/${PROJECT_NAME}/control/sdk/module/dynamixel_sdk.cpp"
-  "src/${PROJECT_NAME}/control/sdk/utils/protocol_1/group_sync_write.cpp"
-  "src/${PROJECT_NAME}/control/sdk/utils/protocol_1/group_bulk_read.cpp"
+  "src/${PROJECT_NAME}/control/sdk/packet/model/group_bulk_read.cpp"
+  "src/${PROJECT_NAME}/control/sdk/packet/protocol_1/group_sync_write.cpp"
+  "src/${PROJECT_NAME}/control/sdk/packet/protocol_1/group_bulk_read.cpp"
+  "src/${PROJECT_NAME}/control/sdk/packet/protocol_2/group_sync_write.cpp"
+  "src/${PROJECT_NAME}/control/sdk/packet/protocol_2/group_bulk_read.cpp"
   "src/${PROJECT_NAME}/imu/node/imu_node.cpp"
   "src/${PROJECT_NAME}/imu/node/imu_provider.cpp"
   "src/${PROJECT_NAME}/joint/model/joint_id.cpp"
